api calls from react

+++ Assignment 3/rpisystem/app/src/main/java/com/nimps/rpisystem/rpisystem/User.java
@@ -15,7 +15,7 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public int id;
+    public long id;
 
     public String first_name;
     public String last_name;
@@ -30,7 +30,7 @@ public class User {
 
     public User(){ }
 
-    public User(int id, String first_name, String last_name, String password, int pantherId){
+    public User(long id, String first_name, String last_name, String password, int pantherId){
         this.id = id;
         this.first_name = first_name;
         this.last_name = last_name;
@@ -46,11 +46,11 @@ public class User {
 
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
